backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Masahiro Sumi",email = "sumi.masahiro312@mai.kyutech.jp"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
